requirements.txt
langchain-openai
langchain-google-genai
google-generativeai